requirements.txt
flask==2.3.3
python-telegram-bot==20.7
requests==2.31.0
gunicorn==21.2.0
python-dotenv==1.0.0